Mainloop/Module: move module library loading/unloading and information maintainance to module.cpp. module.cpp: move module.c to C++. module.cpp: use this to load the static config init function in caerModuleConfigInit(). config_server.cpp: create static config on ADD_MODULE. Remove node fully on REMOVE_MODULE.
