+++ base/module.h
@@ -146,14 +146,36 @@ void caerModuleConfigDefaultListener(sshsNode node, void *userData, enum sshs_no
 		CAER_SYMBOL_EXPORT;
 
 // Functions for mainloop:
-void caerModuleConfigInit(caerModuleFunctions moduleFunctions, sshsNode moduleNode);
+void caerModuleConfigInit(sshsNode moduleNode);
 void caerModuleSM(caerModuleFunctions moduleFunctions, caerModuleData moduleData, size_t memSize,
 	caerEventPacketContainer in, caerEventPacketContainer *out);
-caerModuleData caerModuleInitialize(int16_t moduleID, const char *moduleName, caerModuleFunctions moduleFunctions, sshsNode moduleNode);
+caerModuleData caerModuleInitialize(int16_t moduleID, const char *moduleName, sshsNode moduleNode);
 void caerModuleDestroy(caerModuleData moduleData);
 
 #ifdef __cplusplus
 }
+
+// If Boost version recent enough, use their portable DLL loading support.
+// Else use dlopen() on POSIX systems.
+#include <boost/version.hpp>
+#if defined(BOOST_VERSION) && (BOOST_VERSION / 100000) >= 1 && (BOOST_VERSION / 100 % 1000) >= 61
+#define BOOST_HAS_DLL_LOAD 1
+#else
+#define BOOST_HAS_DLL_LOAD 0
+#endif
+
+#if BOOST_HAS_DLL_LOAD
+#include <boost/dll.hpp>
+using ModuleLibrary = boost::dll::shared_library;
+#else
+#include <dlfcn.h>
+using ModuleLibrary = void *;
+#endif
+
+std::pair<ModuleLibrary, caerModuleInfo> caerLoadModuleLibrary(const std::string &moduleName);
+void caerUnloadModuleLibrary(ModuleLibrary &moduleLibrary);
+void caerUpdateModulesInformation();
+
 #endif
 
 #endif /* MODULE_H_ */
